adding Game States
Here we add the Game States of our Flappy Bird. Game state helps us to organize the different phases of our game.  There are many types of game state like Menu state, Credit state, Actual play state.

In game state, we just need to show a specific area to the users  in the world (eg. Rectangle Area). And we do that thing with a camera where Camera has  a viewport of a specific height and width.