adding Play State
here we defined that when somebody click on the menu state how it should behave. It will transition to the play state.

We rename the States package to State package as the rules of the Package Convention.

Then we use setToOrtho Method to show specific portion of our game world.