Creating Bird Class
In this section, we created a new bird class and used that class to set the position of our bird texture of our screen allowing it to jump up and down.

We implement code for Bird class and set a position for velocity.
A bird actually moved on the screen!